+++ src/collective/nitf/config.py
@@ -200,4 +200,5 @@ JS_RESOURCES = (
     '++resource++collective.js.cycle2/jquery.cycle2.min.js',
     '++resource++collective.js.cycle2/jquery.cycle2.carousel.min.js',
     '++resource++collective.js.cycle2/jquery.cycle2.swipe.min.js',
+    '++resource++collective.js.cycle2/utils.min.js',
 )